+++ apps/bot/src/app/features/bot-config/tabs/ModelBasicInfo.tsx
@@ -3,15 +3,17 @@ import { useNavigate, useParams } from 'react-router-dom';
 import { useQueryClient } from '@tanstack/react-query';
 import { Button, Checkbox, Col, Form, type FormProps, Input, Row } from 'antd';
 import { Log } from '@/log';
-import { confirmModal, toast } from '@/shared-util';
+import { toast } from '@/shared-util';
 import { modelQueryKeys, useDeleteModel, useGetModel, useUpdateModel } from '../hooks/useModelQueries';
 import type { ModelBasicInfoUpdateDatas } from '../types';
 import { FallbackSpinner } from '@/components/custom/FallbackSpinner';
+import { useModal } from '@/libs/shared-ui/src/hooks/useModal';
 
 export default function ModelBasicInfo() {
   const { modelId } = useParams();
   const queryClient = useQueryClient();
   const navigate = useNavigate();
+  const modal = useModal();
   const [form] = Form.useForm();
 
   const { data: model, isFetching } = useGetModel({ params: { modelId } });
@@ -44,7 +46,7 @@ export default function ModelBasicInfo() {
   };
 
   const handleClickDeleteBtn = () => {
-    confirmModal.delete({
+    modal.confirm.delete({
       onOk: () => deleteModel({ modelId }),
     });
   };
